[Feature][UI Next]Added the method of downloading files (#9605)
