+++ dolphinscheduler-ui-next/src/service/service.ts
@@ -93,46 +93,5 @@ service.interceptors.response.use((res: AxiosResponse) => {
   }
 }, err)
 
-const apiPrefix = '/dolphinscheduler'
-const reSlashPrefix = /^\/+/
 
-const resolveURL = (url: string) => {
-  if (url.indexOf('http') === 0) {
-    return url
-  }
-  if (url.charAt(0) !== '/') {
-    return `${apiPrefix}/${url.replace(reSlashPrefix, '')}`
-  }
-
-  return url
-}
-
-/**
- * download file
- */
-const downloadFile = (url: string, obj?: any) => {
-  const param: any = {
-    url: resolveURL(url),
-    obj: obj || {}
-  }
-
-  const form = document.createElement('form')
-  form.action = param.url
-  form.method = 'get'
-  form.style.display = 'none'
-  Object.keys(param.obj).forEach((key) => {
-    const input = document.createElement('input')
-    input.type = 'hidden'
-    input.name = key
-    input.value = param.obj[key]
-    form.appendChild(input)
-  })
-  const button = document.createElement('input')
-  button.type = 'submit'
-  form.appendChild(button)
-  document.body.appendChild(form)
-  form.submit()
-  document.body.removeChild(form)
-}
-
-export { service as axios, downloadFile }
+export { service as axios }
