+++ dolphinscheduler-ui-next/src/utils/index.ts
@@ -19,12 +19,14 @@ import mapping from './mapping'
 import regex from './regex'
 import truncateText from './truncate-text'
 import log from './log'
+import downloadFile from './downloadFile'
 
 const utils = {
   mapping,
   regex,
   truncateText,
-  log
+  log,
+  downloadFile
 }
 
 export default utils
